refactor: Extract CommandLineBuilder into focused modules
Split the monolithic CommandLineBuilder.cs (2,098 lines) into focused,
single-responsibility modules for improved maintainability.

## Quantitative Improvements

- CommandLineBuilder.cs: 2,098 → 175 lines (92% reduction)
- Total code organized across 12 files (2,284 lines)
- Eliminated 3 instances of duplicated helper methods
- 20 magic strings replaced with SectionNames constants

## New Structure

CommandLine/Commands/ (6 files, 1,846 lines)
- ApiCommandDefinitions.cs - type, member, deprecated api
- SearchCommandDefinitions.cs - find, implements, extensions, depends
- UtilityCommandDefinitions.cs - cache, demo, samples, cli, llmstxt, skill, perf
- InspectionCommandDefinitions.cs - diff, library
- RouterCommandDefinition.cs - hidden router command
- PackageCommandDefinitions.cs - package commands

CommandLine/Helpers/ (3 files, 263 lines)
- ArgumentPreprocessor.cs - -NN expansion, implicit command detection
- CommandLineHelpers.cs - shared utilities (file classification, version detection)
- TipWriter.cs - tip generation

Sections/SectionNames.cs - section name constants

## Qualitative Improvements

- Single Responsibility: Each file has one clear purpose
- DRY: Eliminated duplicated TryClassifyAsFilePath, LooksLikeVersionNumber
- Discoverability: Section names centralized as constants
- Testability: Helpers can be unit tested independently
- Maintainability: Adding/modifying commands no longer requires
  navigating a 2,000+ line file
